added buckettreeitem component
