+++ src/components/BucketTreeItemDownloadButton.svelte
