+++ src/components/BucketTreeItemLegend.svelte
