+++ src/components/BucketTreeLabel.svelte
